Refactor backend code. Separated API endpoint routes and initialization code.
