+++ server/.gitignore
@@ -1,1 +1,2 @@
 musicvenv/
+*config.py
